+++ client/src/hooks/useSelectVideo.ts
@@ -1,0 +1,19 @@
+import { useState, useCallback } from 'react';
+import { SelectedVideo } from '@components/Meet/MeetVideo';
+
+export const useSelectVideo = () => {
+  const [selectedVideo, setSelectedVideo] = useState<SelectedVideo | null>(null);
+
+  const deselectVideo = () => setSelectedVideo(null);
+  const selectVideo = useCallback(
+    (videoInfo: SelectedVideo) => () => {
+      setSelectedVideo((selectedVideo) => {
+        if (selectedVideo?.stream?.id === videoInfo?.stream?.id) return null;
+        else return { ...videoInfo };
+      });
+    },
+    [],
+  );
+
+  return { selectVideo, deselectVideo, selectedVideo, setSelectedVideo };
+};
